server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "A Restful API for Tracking Suits"
authors = ["Adrian Van Nierop <adriandvn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-types = "^0.19.1"
django-stubs-ext = "^5.1.1"
ruff = "^0.7.2"
pyright = "^1.1.388"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
extend-exclude = ['**/migrations']

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set.
extend-select = ["E501"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
